Use smallint for enum

+++ db/structure.sql
@@ -16668,7 +16668,7 @@ CREATE TABLE user_statuses (
     emoji character varying DEFAULT 'speech_balloon'::character varying NOT NULL,
     message character varying(100),
     message_html character varying,
-    availability integer
+    availability smallint
 );
 
 CREATE SEQUENCE user_statuses_user_id_seq
